+++ src/WebServer_ESP32_W5500.hpp
@@ -9,12 +9,13 @@
   Built by Khoi Hoang https://github.com/khoih-prog/WebServer_ESP32_W5500
   Licensed under GPLv3 license
 
-  Version: 1.5.2
+  Version: 1.5.3
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.5.1   K Hoang      29/11/2022 Initial coding for ESP32_W5500 (ESP32 + W5500). Sync with WebServer_WT32_ETH01 v1.5.1
   1.5.2   K Hoang      06/01/2023 Suppress compile error when using aggressive compile settings
+  1.5.3   K Hoang      11/01/2023 Using `SPI_DMA_CH_AUTO` and built-in ESP32 MAC
  *****************************************************************************************************************************/
 
 #pragma once
@@ -24,7 +25,7 @@
 
 //////////////////////////////////////////////////////////////
 
-//#define CONFIG_ETH_SPI_ETHERNET_W5500				true
+//#define CONFIG_ETH_SPI_ETHERNET_W5500        true
 
 //////////////////////////////////////////////////////////////
 
@@ -33,12 +34,14 @@
 
 #include <hal/spi_types.h>
 
+//////////////////////////////////////////////////////////////
+
 #if !defined(ETH_SPI_HOST)
   #define ETH_SPI_HOST            SPI3_HOST
 #endif
 
 #if !defined(SPI_CLOCK_MHZ)
-	// Using 25MHz for W5500, 14MHz for W5100
+  // Using 25MHz for W5500, 14MHz for W5100
   #define SPI_CLOCK_MHZ       25
 #endif
 
@@ -62,9 +65,13 @@
   #define CS_GPIO             5
 #endif
 
+//////////////////////////////////////////////////////////////
+
 #ifndef SHIELD_TYPE
   #define SHIELD_TYPE         "ESP32_W5500"
 #endif
+
+//////////////////////////////////////////////////////////////
 
 extern bool ESP32_W5500_eth_connected;